make config static
