+++ react-app/src/constants/config.ts
@@ -1,0 +1,11 @@
+export interface IConfig {
+  API_HOST: string;
+  NODE_ENV: string;
+}
+
+export const getConfig = (): IConfig => {
+  return {
+    API_HOST: '/api/v2',
+    NODE_ENV: import.meta.env.MODE,
+  };
+};
